deleted useless files and changed location of existing functions to util.js
