+++ index.html
@@ -8,7 +8,6 @@
     <link rel="stylesheet" type="text/css" media="screen" href="main.css" />
     
     <script src="/src/gl-matrix-min.js" defer></script>
-    <script type="module" src="/src/scene.js" defer></script>
     <script type="module" src="/src/colors.js" defer></script>
     <script type ="module" src="/src/renderFunctions.js" defer></script>
     <script type ="module" src="/src/input.js" defer></script>
@@ -18,6 +17,7 @@
     <script type ="module" src="/src/shaderProgram.js" defer></script>
     <script type ="module" src="/src/utils.js" defer></script>
     <script type ="module" src="/src/gameState.js" defer></script>
+    <script type="module" src="/src/gameObject.js" defer></script>
     <script type ="module" src="/src/main.js" defer></script>
 
     <script id="vertexShader" type="text">
