+++ src/shapes3d.js
@@ -1,29 +1,5 @@
 import {setCircleVertices, setCircleColor} from './shapes2d.js';
 
-export function rotateObjectMatrixY(object, rotationAngle, translation2){
-    var translation = [ -object[0], -object[1], -object[2]]
-    var matrix = mat4.create();
-    mat4.translate(matrix, matrix,translation2);
-    mat4.rotateY(matrix, matrix,rotationAngle);
-    mat4.translate(matrix, matrix, translation);
-    return matrix;
-}
-
-export function applyTransformation(object, transfMatrix){
-    var len = object.length;
-    var newObject = [];
-    var oldVertex = [];
-    var newVertex = [];
-    for(let i=0; i < len; i += 3){
-        oldVertex[0] = object[i];
-        oldVertex[1] = object[i+1];
-        oldVertex[2] = object[i+2];
-        vec3.transformMat4(newVertex, oldVertex ,transfMatrix);
-        newObject.push(newVertex[0], newVertex[1], newVertex[2]);
-    }
-    return newObject;
-}
-
 export function setLandscapeVertices( len, h){
     const vertexData = [
         // Front
